Allow documented items of the same type to be read from multiple directories
